+++ cli/medperf/web_ui/tests/pages/container/details_page.py
@@ -31,6 +31,26 @@ class ContainerDetailsPage(BasePage):
 
     ACCESS_SECTION = (By.CSS_SELECTOR, "[data-testid='container-access-label']")
 
+    ACCESS_HEADER = (By.CSS_SELECTOR, "h1")
+
+    GRANT_FORM = (By.ID, "grant-access-form")
+    BENCHMARK = (By.ID, "benchmark")
+    EMAIL_INPUT = (By.ID, "email-input")
+    GRANT_BTN = (By.ID, "grant-access-btn")
+
+    BENCHMARK_AUTO = (By.ID, "benchmark-auto")
+    INTERVAL_AUTO = (By.ID, "interval-auto")
+    EMAIL_INPUT_AUTO = (By.ID, "email-input-auto")
+    START_AUTO_BTN = (By.ID, "start-auto-access-btn")
+    STOP_AUTO_BTN = (By.ID, "stop-auto-access-btn")
+    RUNNING_BADGE = (By.ID, "running-badge")
+
+    NO_KEYS_MSG = (By.XPATH, "//p[contains(., 'No users currently have access')]")
+    KEYS_TABLE_ROWS = (By.CSS_SELECTOR, "table tbody tr")
+
+    DELETE_KEYS_FORM = (By.ID, "delete-keys-form")
+    DELETE_KEYS_BTN = (By.ID, "delete-keys-btn")
+
     RESUME_SCRIPT = (
         By.XPATH,
         "//script[not(@src)][contains(., 'resumeRunningTask')]",
@@ -68,3 +88,25 @@ class ContainerDetailsPage(BasePage):
         associations = self.find(self.ASSOCIATIONS_LIST)
         self.wait_for_visibility_element(associations)
         return [i.text for i in self.driver.find_elements(*self.ASSOCIATION_CARDS)]
+
+    def revoke_btn(self, key_id):
+        return (By.ID, f"revoke-btn-{key_id}")
+
+    def grant_access(self, benchmark, emails):
+        self.select_searchable_entity(self.BENCHMARK, benchmark)
+        self.type(self.EMAIL_INPUT, ",".join(emails) + ",")
+        self.click(self.GRANT_BTN)
+
+    def revoke_access(self, key_id):
+        self.click(self.revoke_btn(key_id))
+
+    def delete_keys(self):
+        self.click(self.DELETE_KEYS_BTN)
+
+    def start_auto_access(self, benchmark, emails):
+        self.select_searchable_entity(self.BENCHMARK_AUTO, benchmark)
+        self.type(self.EMAIL_INPUT_AUTO, ",".join(emails) + ",")
+        self.click(self.START_AUTO_BTN)
+
+    def stop_auto_access(self):
+        self.click(self.STOP_AUTO_BTN)
